requirements.txt
opencv-python
rembg
pillow
